requirements.txt
annotated-types==0.7.0
anyio==4.4.0
APScheduler==3.10.4
certifi==2024.7.4
distro==1.9.0
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
openai==1.36.1
pydantic==2.8.2
pydantic_core==2.20.1
python-dotenv==1.0.1
python-telegram-bot==21.4
pytz==2024.1
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.31
tqdm==4.66.4
typing_extensions==4.12.2
tzlocal==5.2
